modified discount field on product table and added new endopoint for product create

+++ database/migrations/2019_06_13_1_create_products_table.php
@@ -20,7 +20,7 @@ class CreateProductsTable extends Migration
             $table->text('detail');
             $table->decimal('price');
             $table->integer('stock');
-            $table->decimal('discount');
+            $table->integer('discount');
             $table->timestamps();
 
         });
